Display 3 months of shifts and clean up layout.
