+++ app/javascript/stylesheets/volunteer.scss
@@ -1,4 +1,6 @@
 .shift-schedule .calendar-lg {
+  margin-bottom: 2rem;
+
   .calendar-events {
     line-height: 1.2;
   }
@@ -12,6 +14,17 @@
     .date-item {
       pointer-events: none;
     }
+  }
+  .calendar-date.prev-month,
+  .calendar-date.next-month {
+    opacity: 0.7;
+  }
+  .calendar-nav {
+    justify-content: center;
+  }
+  
+  .short-staffed {
+    font-weight: bold;
   }
 }
