+++ app/views/volunteer/shifts/index.html.erb
@@ -17,57 +17,19 @@
     <% end %>
   </div>
 </div>
-<div class="columns">
-  <div class="column col-12 col-md-10 col-sm-12">
-    <div class="shift-schedule">
 
-<div class="calendar calendar-lg">
-  <div class="calendar-nav navbar">
-    <button class="btn btn-action btn-link btn-lg">
-      <i class="icon icon-arrow-left"></i>
-    </button>
-    <div class="navbar-primary"><%= @month_calendar.title %></div>
-    <button class="btn btn-action btn-link btn-lg">
-      <i class="icon icon-arrow-right"></i>
-    </button>
-  </div>
-
-  <div class="calendar-container">
-    <div class="calendar-header">
-      <div class="calendar-date">Sun</div>
-      <div class="calendar-date">Mon</div>
-      <div class="calendar-date">Tue</div>
-      <div class="calendar-date">Wed</div>
-      <div class="calendar-date">Thu</div>
-      <div class="calendar-date">Fri</div>
-      <div class="calendar-date">Sat</div>
+<% @month_calendars.each do |month_calendar| %>
+  <div class="columns">
+    <div class="column col-12 col-md-10 col-sm-12">
+      <div class="shift-schedule">
+        <%= render partial: "month_calendar", locals: { month_calendar: month_calendar, attendee: @attendee } %>
+      </div>
     </div>
-
-    <div class="calendar-body">
-      <% @month_calendar.each_day do |day| %>
-        <div class="calendar-date <%= calendar_day_class(day) %>">
-          <span class="date-item <%= calendar_date_item_class(day) %>" <%= "disabled" unless day.events? %>><%= day.number %></span>
-          <div class="calendar-events">
-            <% day.each_shift do |shift| %>
-              <% if @attendee && shift.attended_by?(@attendee.email) %>
-                <%= link_to volunteer_event_path(event_ids: shift.event_ids), class: "calendar-event bg-primary text-light", title: "You are signed up for this shift." do %>
-                  <%= shift.times %><br><%= shift.title %>
-                <% end %>
-              <% else %>
-                <%= link_to volunteer_event_path(event_ids: shift.event_ids), class: "calendar-event" do %>
-                  <%= shift.times %><br><%= shift.title %>
-                <% end %>
-              <% end %>
-            <% end %>
-          </div>
-        </div>
-      <% end %>
-    </div>
-
-    <% unless signed_in_via_google? %>
-      <p class="text-center mt-2">
-        <%= link_to "Sign in with Google", "/auth/google_oauth2", method: :post %>
-      </p>
-    <% end %>
   </div>
-</div>
+<% end %>
+
+<% unless signed_in_via_google? %>
+  <p class="text-center mt-2">
+    <%= link_to "Sign in with Google", "/auth/google_oauth2", method: :post %>
+  </p>
+<% end %>
